# API_documentacion.yaml
openapi: 3.0.1
info:
  title: Delilah resto API
  description: 'Servicios REST para los distintos productos y usuarios'
  contact:
    email: morenodanielaeliana@gmail.com
  version: 1.0.0
...
paths:
  /order:
    post:
      tags:
      - order
      summary: Agregar un nuevo pedido
      requestBody:
        description: Detalle del pedido que se desea agregar. Usuario con rol de usuario puede agregar un pedido
        content:
          application/json:
            schema:
             type: object
             properties:
              usernameId: 
                type: string
                example: "5ed2b31fcf481e34185ddf85"
              productsId: 
                type: array
                items:
                  example:
                  - "5ed2b31fcf481e34185ddf85"
                  - "5ed2b7edab3fcc2b48a73bb3"
              payment_methodId: 
                type: string
                example: "5ed2c09aab3fcc2b48a73bc8"
              delivery_address: 
                type: string
                example: "Avellaneda 1235"
        required: true
      responses:
        200:
          description: order added sucessfully
          content: {}
        400:
          description: invalid properties
          content: {}
        401:
          description: user not authorized
          content: {}
      security:
      - bearerAuth: []
      x-codegen-request-body-name: body
    get:
      tags:
      - order
      summary: Obtener todos los pedidos
      description: Usuario con rol de administrador puede obtener todos los pedidos
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Orders'
        401:
          description: user not authorized
          content: {}
      security:
      - bearerAuth: []
  /order/{id}:
    patch:
      tags:
      - order
      summary: Modificar un pedido
      parameters:
      - name: id
        in: path
        description: ID del pedido a modificar
        required: true
        schema:
          type: string
      requestBody:
        description: Objecto con el ID del status correspondiente se quiere modificar. Usuario con rol de administrador puede modificar un pedido
        content:
          application/json:
            schema:
             type: object
             properties:
              orderStatusId: 
                type: string
                example: "5ecae9e330c1b02f10cd7acb"
        required: true
      responses:
        200:
          description: order status updated sucessfully
          content: {}
        400:
          description: invalid properties
          content: {}
        401:
          description: user not authorized
          content: {}
      security:
      - bearerAuth: []
    delete:
      tags:
      - order
      summary: Eliminar un pedido por ID
      description: Usuario con rol de administrador puede eliminar un pedido.
      parameters:
      - name: id
        in: path
        description: ID del pedido que se desea eliminar
        required: true
        schema:
          type: string
      responses:
        200:
          description: order deleted sucessfully
          content: {}
        401:
          description: user not authorized
          content: {}
      security:
      - bearerAuth: []
    get:
      tags:
      - order
      summary: Obtener un pedido por ID
      description: Usuario con rol de administrar puede obtener un pedido por ID
      parameters:
      - name: "id"
        in: "path"
        schema:
          type: string
        required: true
        description: ID del pedido que se quiere obtener
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref:  '#/components/schemas/Order'
        401:
          description: user not authorized
          content: {}
      security:
      - bearerAuth: []
  /myorder: 
    get:
      tags:
      - order
      summary: Obtener el pedido del usuario
      description: El usuario con rol de usuario solo podrá ver los pedidos realizados por si mismo.
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/MyOrders'
        401:
          description: user not authorized
          content: {}
      security:
      - bearerAuth: []
  /myorder/{id}: 
    get:
      tags:
      - order
      summary: Obtener el pedido del usuario por ID
      description: Usuario con rol de usuario puede obtener un pedido por ID realizado por si mismo.
      parameters:
      - name: id
        in: path
        description: ID del pedido que se quiere obtener
        required: true
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/MyOrder'
        401:
          description: user not authorized
          content: {}
  /order/status:
    post:
      tags:
      - order status
      summary: Agregar un nuevo estado de un pedido
      requestBody:
        description: Detalle del estado que se desea agregar. Usuario con rol de administrador puede agregar un nuevo estado.
        content:
          application/json:
            schema:
             type: object
             properties:
              description: 
                type: string
                example: "Nuevo"
        required: true
      responses:
        200:
          description: order status added sucessfully
          content: {}
        400:
          description: invalid properties
          content: {}
        401:
          description: user not authorized
          content: {}
      security:
      - bearerAuth: []
    get:
      tags:
      - order status
      summary: Obtener todos los estados de pedidos
      description: Usuario con rol de administrador puede obtener todos los estados de pedidos
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderStatuses'
        401:
          description: user not authorized
          content: {}
      security:
      - bearerAuth: []
  /order/status/{id}:
    patch:
      tags:
      - order status
      summary: Modificar un estado de pedido
      parameters:
      - name: id
        in: path
        description: ID del estado de pedido a modificar
        required: true
        schema:
          type: string
      requestBody:
        description: Objecto con el ID del estado del pedido correspondiente se quiere modificar. Usuario con rol de administrador puede modificar un estado de pedido
        content:
          application/json:
            schema:
             type: object
             properties:
              description: 
                type: string
                example: "En camino"
        required: true
      responses:
        200:
          description: order status updated sucessfully
          content: {}
        400:
          description: invalid properties
          content: {}
        401:
          description: user not authorized
          content: {}
      security:
      - bearerAuth: []
    delete:
      tags:
      - order status
      summary: Eliminar un estado de pedido por ID
      description: Usuario con rol de administrador puede eliminar un estado de pedido.
      parameters:
      - name: id
        in: path
        description: ID del estado de pedido que se desea eliminar
        required: true
        schema:
          type: string
      responses:
        200:
          description: order status deleted sucessfully
          content: {}
        401:
          description: user not authorized
          content: {}
      security:
      - bearerAuth: []
    get:
      tags:
      - order status
      summary: Obtener un estado de pedido por ID
      description: Usuario con rol de administrar puede obtener un estado de pedido por ID
      parameters:
      - name: "id"
        in: "path"
        schema:
          type: string
        required: true
        description: ID del estado de pedido que se quiere obtener
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref:  '#/components/schemas/OrderStatus'
        401:
          description: user not authorized
          content: {}
      security:
      - bearerAuth: []
...
components:
  schemas:
    Order: 
      type: object
      properties:
        _id: 
          $ref: '#/components/schemas/orderId'
        username:
          $ref: '#/components/schemas/user_username'
        products_description:
          type: array
          items:
            $ref: '#/components/schemas/product_description'
        payment_description:
          $ref: '#/components/schemas/payment_method_description'
        name_lastname:
          $ref: '#/components/schemas/user_name_lastname' 
        status_description:
          $ref: '#/components/schemas/order_status_description'
        total:
          $ref: '#/components/schemas/total'   
        delivery_address:
          $ref: '#/components/schemas/user_delivery_address'
        user_email:
          $ref: '#/components/schemas/user_email'
        user_telephone:
          $ref: '#/components/schemas/user_telephone'
        user_address:
          $ref: '#/components/schemas/user_address'
    Orders:  
      type: array
      items:
        $ref: '#/components/schemas/Order'
    orderId: 
      type: string
      example: '5ed2cbddab3fcc2b48a73bd1'
    total: 
      type: string
      example: '250' 
    product_description: 
      type: string
      example: Focaccia veggie
...
    user_username: 
      type: string
      example: 'daniela18'
    user_delivery_address: 
      type: string
      example: 'Avellaneda 1200'
    user_email: 
      type: string
      example: 'damoreno@gmail.com'
    user_name_lastname: 
      type: string
      example: "Daniela Moreno"
    user_telephone: 
      type: string
      example: '1158749632'
    user_address: 
      type: string
      example: 'Colón 1343'
...
    MyOrder: 
      type: object
      properties:
        _id: 
          $ref: '#/components/schemas/orderId'
        products_description:
          type: array
          items:
            $ref: '#/components/schemas/product_description'
        payment_description:
          $ref: '#/components/schemas/payment_method_description'
        status_description:
          $ref: '#/components/schemas/order_status_description'
        total:
          $ref: '#/components/schemas/total'   
        delivery_address:
          $ref: '#/components/schemas/user_delivery_address'
    MyOrders: 
      type: array
      items:
        $ref: '#/components/schemas/MyOrder'
...
    order_status_description:
      type: string
      example: "En camino"
    order_status_id: 
      type: string
      example: "5ecae9f530c1b02f10cd7acd"
    OrderStatus: 
      type: object
      properties: 
        _id:
          $ref: '#/components/schemas/order_status_id'
        description:
          $ref: '#/components/schemas/order_status_description'
    OrderStatuses:
      type: array
      items:
        $ref: '#/components/schemas/OrderStatus'
    payment_method_description:
      type: string
      example: "Efectivo"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT